--- src/components/Header.js
@@ -1,4 +1,0 @@
-import { Container } from "react-bootstrap";
-
-function Header() {}
-export default Header;
