+++ src/components/logSign/LogSign.js
@@ -1,0 +1,14 @@
+function LogSign() {
+  // position relative a header
+  console.log("renderizado logIn");
+
+  return (
+    <div className="d-flex gap-2 border">
+      <button>LOGIN</button>
+      <p className="text light"> / /</p>
+      <button>SIGNUP</button>
+    </div>
+  );
+}
+
+export default LogSign;
